Refactor large modules into focused submodules
