+++ custom_components/thessla_green_modbus/modbus_transport.py
@@ -1,833 +1,35 @@
 """Transport abstractions for Modbus communication."""
 
-from __future__ import annotations
-
-import asyncio
-import inspect
-import logging
-from abc import ABC, abstractmethod
-from typing import Any
-
-from pymodbus.client import AsyncModbusTcpClient
-
-try:  # pragma: no cover
-    from pymodbus.client import AsyncModbusSerialClient as _AsyncModbusSerialClient
-except (ImportError, AttributeError) as serial_import_err:  # pragma: no cover
-    _AsyncModbusSerialClient = None
-    SERIAL_IMPORT_ERROR: Exception | None = serial_import_err
-else:  # pragma: no cover
-    SERIAL_IMPORT_ERROR = None
-
-from ._transport_retry import apply_transport_backoff, log_transport_retry
-from .const import CONNECTION_TYPE_TCP, CONNECTION_TYPE_TCP_RTU
-from .error_contract import classify_error
-from .error_policy import to_log_message
-from .modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
-from .modbus_helpers import (
-    _call_modbus,
-    async_maybe_await_close,
-    get_rtu_framer,
+from .modbus_helpers import get_rtu_framer
+from .modbus_transport_base import (
+    BaseModbusTransport,
+    RawModbusResponse,
+    RawModbusWriteResponse,
+    _append_crc,
+    _crc16,
+    classify_transport_error,
 )
-
-_LOGGER = logging.getLogger(__name__)
-_MIN_SLAVE_ID = 1
-_MAX_SLAVE_ID = 247
-_MAX_READ_REGISTERS = 125
-_MAX_WRITE_REGISTERS = 123
-
-
-def classify_transport_error(exc: BaseException) -> tuple[str, str]:
-    """Expose normalized retry classification for transport layer tests."""
-
-    contract = classify_error(exc)
-    return contract.kind, contract.reason
-
-
-def _crc16(data: bytes) -> int:
-    """Return Modbus RTU CRC16 for the provided payload."""
-
-    crc = 0xFFFF
-    for byte in data:
-        crc ^= byte
-        for _ in range(8):
-            if crc & 1:
-                crc = (crc >> 1) ^ 0xA001
-            else:
-                crc >>= 1
-    return crc & 0xFFFF
-
-
-def _append_crc(data: bytes) -> bytes:
-    """Append Modbus RTU CRC16 in little-endian order."""
-
-    return data + _crc16(data).to_bytes(2, "little")
-
-
-class RawModbusResponse:
-    """Minimal Modbus response container for raw RTU-over-TCP reads."""
-
-    def __init__(self, registers: list[int] | None = None) -> None:
-        self.registers = registers or []
-
-    def isError(self) -> bool:
-        return False
-
-
-class RawModbusWriteResponse:
-    """Minimal Modbus response container for raw RTU-over-TCP writes."""
-
-    def isError(self) -> bool:
-        return False
-
-
-class BaseModbusTransport(ABC):
-    """Base interface for Modbus transports."""
-
-    def __init__(
-        self,
-        *,
-        max_retries: int,
-        base_backoff: float,
-        max_backoff: float,
-        timeout: float,
-        offline_state: bool = False,
-    ) -> None:
-        self.max_retries = max(1, int(max_retries))
-        self.base_backoff = max(0.0, float(base_backoff))
-        self.max_backoff = max(0.0, float(max_backoff))
-        self.timeout = float(timeout)
-        self.offline_state = offline_state
-        self._lock = asyncio.Lock()
-
-    @property
-    def offline(self) -> bool:
-        """Return whether the transport is offline."""
-
-        return self.offline_state
-
-    def is_connected(self) -> bool:
-        """Return whether the transport is currently connected."""
-
-        return self._is_connected()
-
-    async def _execute(self, func: Any, *, ensure_connection: bool = False) -> Any:
-        """Execute a transport operation with shared error handling."""
-
-        try:
-            if ensure_connection:
-                await self.ensure_connected()
-            response = await func()
-            self.offline_state = False
-            return response
-        except TimeoutError:
-            self.offline_state = True
-            await self._reset_connection()
-            raise
-        except ModbusIOException:
-            self.offline_state = True
-            await self._reset_connection()
-            raise
-        except (ConnectionException, OSError):
-            self.offline_state = True
-            await self._reset_connection()
-            raise
-        except asyncio.CancelledError:
-            self.offline_state = True
-            try:
-                await self._reset_connection()
-            except (ConnectionException, ModbusException, OSError, RuntimeError) as exc:
-                _LOGGER.debug("Reset connection failed during CancelledError handling: %s", exc)
-            raise
-        except ModbusException as exc:
-            kind, reason = classify_transport_error(exc)
-            _LOGGER.error("Permanent Modbus error (%s/%s): %s", kind, reason, to_log_message(exc))
-            self.offline_state = True
-            raise
-        except (
-            AttributeError,
-            RuntimeError,
-            TypeError,
-            ValueError,
-        ) as exc:  # pragma: no cover - unexpected
-            _LOGGER.error("Unexpected transport error: %s", to_log_message(exc))
-            self.offline_state = True
-            raise
-
-    async def call(
-        self,
-        func: Any,
-        slave_id: int,
-        *args: Any,
-        attempt: int = 1,
-        max_attempts: int | None = None,
-        backoff: float | None = None,
-        backoff_jitter: float | tuple[float, float] | None = None,
-        apply_backoff: bool = True,
-        **kwargs: Any,
-    ) -> Any:
-        """Call a Modbus function with connection management and retries."""
-
-        async def _invoke() -> Any:
-            return await _call_modbus(
-                func,
-                slave_id,
-                *args,
-                attempt=attempt,
-                max_attempts=max_attempts or self.max_retries,
-                timeout=self.timeout,
-                backoff=backoff if backoff is not None else self.base_backoff,
-                backoff_jitter=backoff_jitter,
-                apply_backoff=apply_backoff,
-                **kwargs,
-            )
-
-        return await self._execute(_invoke, ensure_connection=True)
-
-    async def ensure_connected(self) -> None:
-        """Ensure the underlying transport is connected."""
-
-        async with self._lock:
-            if self._is_connected():
-                return
-            await self._reset_connection()
-            await self._connect()
-
-    async def close(self) -> None:
-        """Close the transport."""
-
-        async with self._lock:
-            await self._reset_connection()
-            self.offline_state = True
-
-    async def _handle_timeout(self, attempt: int, exc: Exception) -> None:
-        """Handle timeout errors with reconnection and backoff."""
-
-        log_transport_retry(
-            logger=_LOGGER,
-            operation="timeout",
-            attempt=attempt,
-            max_attempts=self.max_retries,
-            exc=exc,
-            base_backoff=self.base_backoff,
-        )
-        self.offline_state = True
-        await self._reset_connection()
-        await self._apply_backoff(attempt)
-
-    async def _handle_transient(self, attempt: int, exc: Exception) -> None:
-        """Handle transient transport errors."""
-
-        log_transport_retry(
-            logger=_LOGGER,
-            operation="transient",
-            attempt=attempt,
-            max_attempts=self.max_retries,
-            exc=exc,
-            base_backoff=self.base_backoff,
-        )
-        self.offline_state = True
-        await self._reset_connection()
-        await self._apply_backoff(attempt)
-
-    async def _apply_backoff(self, attempt: int) -> None:
-        """Sleep for the calculated backoff duration respecting the maximum."""
-        await apply_transport_backoff(
-            attempt=attempt, base_backoff=self.base_backoff, max_backoff=self.max_backoff
-        )
-
-    @abstractmethod
-    def _is_connected(self) -> bool:
-        """Return whether the underlying client is connected."""
-
-    @abstractmethod
-    async def _connect(self) -> None:
-        """Connect the underlying transport."""
-
-    @abstractmethod
-    async def _reset_connection(self) -> None:
-        """Reset the underlying connection."""
-
-    @abstractmethod
-    async def read_input_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        """Read input registers from the device."""
-
-    @abstractmethod
-    async def read_holding_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        """Read holding registers from the device."""
-
-    @abstractmethod
-    async def write_register(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        value: int,
-        attempt: int = 1,
-    ) -> Any:
-        """Write a single holding register."""
-
-    @abstractmethod
-    async def write_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        values: list[int],
-        attempt: int = 1,
-    ) -> Any:
-        """Write multiple holding registers."""
-
-
-class _ClientBackedTransport(BaseModbusTransport):
-    """Shared implementation for transports backed by a pymodbus-like client."""
-
-    client: Any | None = None
-
-    async def _ensure_client(self) -> Any:
-        if self.client is None:
-            await self.ensure_connected()
-        assert self.client is not None
-        return self.client
-
-    async def _invoke_client(
-        self, method_name: str, slave_id: int, address: int, **kwargs: Any
-    ) -> Any:
-        client = await self._ensure_client()
-        func = getattr(client, method_name)
-        return await self.call(func, slave_id, address, **kwargs)
-
-    async def _connect_client(self, *, endpoint: str) -> None:
-        client = self.client
-        if client is None:  # pragma: no cover
-            raise ConnectionException("Internal error: client not initialized before connect")
-        connect_method = getattr(client, "connect", None)
-        if callable(connect_method):
-            connected = connect_method()
-            if inspect.isawaitable(connected):
-                connected = await connected
-        else:
-            connected = True
-            client.connected = True
-        if not connected:
-            self.offline_state = True
-            raise ConnectionException(f"Could not connect to {endpoint}")
-        self.offline_state = False
-
-    async def _reset_connection(self) -> None:
-        client = self.client
-        if client is None:
-            return
-        try:
-            await async_maybe_await_close(client)
-        finally:
-            self.client = None
-
-    async def read_input_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        return await self._invoke_client(
-            "read_input_registers",
-            slave_id,
-            address,
-            count=count,
-            attempt=attempt,
-        )
-
-    async def read_holding_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        return await self._invoke_client(
-            "read_holding_registers",
-            slave_id,
-            address,
-            count=count,
-            attempt=attempt,
-        )
-
-    async def write_register(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        value: int,
-        attempt: int = 1,
-    ) -> Any:
-        return await self._invoke_client(
-            "write_register",
-            slave_id,
-            address,
-            value=value,
-            attempt=attempt,
-        )
-
-    async def write_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        values: list[int],
-        attempt: int = 1,
-    ) -> Any:
-        return await self._invoke_client(
-            "write_registers",
-            slave_id,
-            address,
-            values=values,
-            attempt=attempt,
-        )
-
-
-class TcpModbusTransport(_ClientBackedTransport):
-    """TCP Modbus transport implementation."""
-
-    def __init__(
-        self,
-        *,
-        host: str,
-        port: int,
-        connection_type: str = CONNECTION_TYPE_TCP,
-        max_retries: int,
-        base_backoff: float,
-        max_backoff: float,
-        timeout: float,
-        offline_state: bool = False,
-    ) -> None:
-        super().__init__(
-            max_retries=max_retries,
-            base_backoff=base_backoff,
-            max_backoff=max_backoff,
-            timeout=timeout,
-            offline_state=offline_state,
-        )
-        self.host = host
-        self.port = port
-        self.connection_type = connection_type
-        self.client: AsyncModbusTcpClient | None = None
-
-    def _is_connected(self) -> bool:
-        return bool(self.client and getattr(self.client, "connected", False))
-
-    def _build_tcp_client(self, *, framer: Any | None = None) -> AsyncModbusTcpClient:
-        from pymodbus.client import AsyncModbusTcpClient as AsyncTcpClient
-
-        common_kwargs: dict[str, Any] = {
-            "port": self.port,
-            "timeout": self.timeout,
-            "reconnect_delay": 1,
-            "reconnect_delay_max": 300,
-            "retries": self.max_retries,
-        }
-        if framer is not None:
-            common_kwargs["framer"] = framer
-
-        try:
-            return AsyncTcpClient(self.host, **common_kwargs)
-        except TypeError as exc:
-            _LOGGER.debug(
-                "AsyncModbusTcpClient does not accept reconnect parameters: %s",
-                exc,
-            )
-            fallback_kwargs = {k: v for k, v in common_kwargs.items() if k in {"port", "timeout"}}
-            if framer is not None:
-                fallback_kwargs["framer"] = framer
-            try:
-                return AsyncTcpClient(self.host, **fallback_kwargs)
-            except TypeError:
-                client = AsyncTcpClient()
-                client.host = self.host
-                client.port = self.port
-                return client
-
-    async def _connect(self) -> None:
-        if self.connection_type == CONNECTION_TYPE_TCP_RTU:
-            framer = get_rtu_framer()
-            if framer is None:
-                message = (
-                    "RTU over TCP requires pymodbus with RTU framer support "
-                    "(FramerType.RTU or ModbusRtuFramer)."
-                )
-                _LOGGER.error(message)
-                raise ConnectionException(message)
-            _LOGGER.info(
-                "Connecting Modbus TCP RTU to %s:%s (timeout=%s)",
-                self.host,
-                self.port,
-                self.timeout,
-            )
-            try:
-                self.client = self._build_tcp_client(framer=framer)
-            except TypeError as exc:
-                message = (
-                    "RTU over TCP is not supported by the installed pymodbus client. "
-                    "Please upgrade pymodbus."
-                )
-                _LOGGER.error("%s (%s)", message, exc)
-                raise ConnectionException(message) from exc
-        else:
-            _LOGGER.info(
-                "Connecting Modbus TCP to %s:%s (timeout=%s)",
-                self.host,
-                self.port,
-                self.timeout,
-            )
-            self.client = self._build_tcp_client()
-        await self._connect_client(endpoint=f"{self.host}:{self.port}")
-        _LOGGER.debug("TCP Modbus connection established to %s:%s", self.host, self.port)
-
-
-class RtuModbusTransport(_ClientBackedTransport):
-    """RTU Modbus transport implementation using async serial client."""
-
-    def __init__(
-        self,
-        *,
-        serial_port: str,
-        baudrate: int,
-        parity: str,
-        stopbits: int,
-        max_retries: int,
-        base_backoff: float,
-        max_backoff: float,
-        timeout: float,
-        offline_state: bool = False,
-    ) -> None:
-        super().__init__(
-            max_retries=max_retries,
-            base_backoff=base_backoff,
-            max_backoff=max_backoff,
-            timeout=timeout,
-            offline_state=offline_state,
-        )
-        self.serial_port = serial_port
-        self.baudrate = baudrate
-        self.parity = parity
-        self.stopbits = stopbits
-        self.client: _AsyncModbusSerialClient | None = None
-
-    def _is_connected(self) -> bool:
-        return bool(self.client and getattr(self.client, "connected", False))
-
-    async def _connect(self) -> None:
-        if _AsyncModbusSerialClient is None:
-            message = "Modbus serial client is unavailable. Install pymodbus with serial support."
-            if SERIAL_IMPORT_ERROR is not None:
-                message = f"{message} ({SERIAL_IMPORT_ERROR})"
-            raise ConnectionException(message)
-        if not self.serial_port:
-            raise ConnectionException("Serial port not configured for RTU transport")
-        self.client = _AsyncModbusSerialClient(
-            method="rtu",
-            port=self.serial_port,
-            baudrate=self.baudrate,
-            parity=self.parity,
-            stopbits=self.stopbits,
-            timeout=self.timeout,
-        )
-        await self._connect_client(endpoint=self.serial_port)
-        _LOGGER.debug("RTU Modbus connection established on %s", self.serial_port)
-
-
-class RawRtuOverTcpTransport(BaseModbusTransport):
-    """RTU-over-TCP transport that sends raw Modbus RTU frames over TCP."""
-
-    def __init__(
-        self,
-        *,
-        host: str,
-        port: int,
-        max_retries: int,
-        base_backoff: float,
-        max_backoff: float,
-        timeout: float,
-        offline_state: bool = False,
-    ) -> None:
-        super().__init__(
-            max_retries=max_retries,
-            base_backoff=base_backoff,
-            max_backoff=max_backoff,
-            timeout=timeout,
-            offline_state=offline_state,
-        )
-        self.host = host
-        self.port = port
-        self._reader: asyncio.StreamReader | None = None
-        self._writer: asyncio.StreamWriter | None = None
-        self._request_lock = asyncio.Lock()
-
-    def _is_connected(self) -> bool:
-        return bool(self._writer and not self._writer.is_closing())
-
-    async def _connect(self) -> None:
-        try:
-            self._reader, self._writer = await asyncio.wait_for(
-                asyncio.open_connection(self.host, self.port),
-                timeout=self.timeout,
-            )
-        except TimeoutError as exc:
-            raise TimeoutError(f"Timed out connecting to {self.host}:{self.port}") from exc
-        except OSError as exc:
-            raise ConnectionException(f"Could not connect to {self.host}:{self.port}") from exc
-
-    async def _reset_connection(self) -> None:
-        if self._writer is None:
-            self._reader = None
-            return
-        try:
-            self._writer.close()
-            wait_closed = getattr(self._writer, "wait_closed", None)
-            if callable(wait_closed):
-                await wait_closed()
-        finally:
-            self._reader = None
-            self._writer = None
-
-    async def _read_exactly(self, size: int) -> bytes:
-        if self._reader is None:
-            raise ConnectionException("RTU-over-TCP socket not connected")
-        try:
-            return await asyncio.wait_for(self._reader.readexactly(size), timeout=self.timeout)
-        except asyncio.IncompleteReadError as exc:
-            raise ModbusIOException("Incomplete RTU response") from exc
-        except TimeoutError as exc:
-            raise TimeoutError("Timed out waiting for RTU response") from exc
-
-    @staticmethod
-    def _validate_crc(payload: bytes, crc_bytes: bytes) -> None:
-        expected = _crc16(payload).to_bytes(2, "little")
-        if crc_bytes != expected:
-            raise ModbusIOException("CRC mismatch in RTU response")
-
-    @staticmethod
-    def _build_read_frame(slave_id: int, function: int, address: int, count: int) -> bytes:
-        payload = bytes(
-            [
-                slave_id & 0xFF,
-                function & 0xFF,
-                (address >> 8) & 0xFF,
-                address & 0xFF,
-                (count >> 8) & 0xFF,
-                count & 0xFF,
-            ]
-        )
-        return _append_crc(payload)
-
-    @staticmethod
-    def _validate_slave_id(slave_id: int) -> None:
-        if not (_MIN_SLAVE_ID <= slave_id <= _MAX_SLAVE_ID):
-            raise ModbusIOException(
-                f"Invalid slave_id={slave_id}; expected {_MIN_SLAVE_ID}-{_MAX_SLAVE_ID}"
-            )
-
-    @staticmethod
-    def _validate_read_count(count: int) -> None:
-        if not (1 <= count <= _MAX_READ_REGISTERS):
-            raise ModbusIOException(f"Invalid read count={count}; expected 1-{_MAX_READ_REGISTERS}")
-
-    @staticmethod
-    def _validate_write_count(qty: int) -> None:
-        if not (1 <= qty <= _MAX_WRITE_REGISTERS):
-            raise ModbusIOException(
-                f"Invalid write quantity={qty}; expected 1-{_MAX_WRITE_REGISTERS}"
-            )
-
-    @staticmethod
-    def _build_write_single_frame(slave_id: int, address: int, value: int) -> bytes:
-        payload = bytes(
-            [
-                slave_id & 0xFF,
-                6,
-                (address >> 8) & 0xFF,
-                address & 0xFF,
-                (value >> 8) & 0xFF,
-                value & 0xFF,
-            ]
-        )
-        return _append_crc(payload)
-
-    @staticmethod
-    def _build_write_multiple_frame(slave_id: int, address: int, values: list[int]) -> bytes:
-        qty = len(values)
-        payload = bytearray(
-            [
-                slave_id & 0xFF,
-                16,
-                (address >> 8) & 0xFF,
-                address & 0xFF,
-                (qty >> 8) & 0xFF,
-                qty & 0xFF,
-                qty * 2,
-            ]
-        )
-        for value in values:
-            payload.extend([(value >> 8) & 0xFF, value & 0xFF])
-        return _append_crc(bytes(payload))
-
-    async def _read_response(self, slave_id: int, function: int) -> bytes:
-        header = await self._read_exactly(2)
-        resp_slave, resp_func = header
-
-        if resp_slave != (slave_id & 0xFF):
-            raise ModbusIOException("Unexpected slave ID in RTU response")
-
-        if resp_func & 0x80:
-            exception_code = await self._read_exactly(1)
-            crc_bytes = await self._read_exactly(2)
-            payload = header + exception_code
-            self._validate_crc(payload, crc_bytes)
-            raise ModbusException(f"Modbus exception {exception_code[0]} for function {resp_func}")
-
-        if resp_func != (function & 0xFF):
-            raise ModbusIOException("Unexpected function code in RTU response")
-
-        if function in (3, 4):
-            byte_count_raw = await self._read_exactly(1)
-            byte_count = byte_count_raw[0]
-            data = await self._read_exactly(byte_count)
-            crc_bytes = await self._read_exactly(2)
-            payload = header + byte_count_raw + data
-            self._validate_crc(payload, crc_bytes)
-            return data
-
-        body = await self._read_exactly(4)
-        crc_bytes = await self._read_exactly(2)
-        payload = header + body
-        self._validate_crc(payload, crc_bytes)
-        return body
-
-    async def _send_frame(self, frame: bytes, slave_id: int, function: int) -> bytes:
-        async with self._request_lock:
-            if self._writer is None:
-                raise ConnectionException("RTU-over-TCP socket not connected")
-            self._writer.write(frame)
-            await self._writer.drain()
-            return await self._read_response(slave_id, function)
-
-    @staticmethod
-    def _decode_register_words(data: bytes, *, count: int) -> list[int]:
-        expected_bytes = count * 2
-        if len(data) != expected_bytes:
-            raise ModbusIOException("Invalid byte count in RTU response")
-        return [int.from_bytes(data[i : i + 2], "big") for i in range(0, len(data), 2)]
-
-    @staticmethod
-    def _validate_write_echo(response: bytes, *, address: int, expected_value: int) -> None:
-        if len(response) != 4:
-            raise ModbusIOException("Invalid write response length")
-        resp_addr = (response[0] << 8) | response[1]
-        resp_value = (response[2] << 8) | response[3]
-        if resp_addr != address or resp_value != expected_value:
-            raise ModbusIOException("Write response does not match request")
-
-    async def _read_registers_common(
-        self, *, slave_id: int, address: int, count: int, function: int
-    ) -> RawModbusResponse:
-        frame = self._build_read_frame(slave_id, function, address, count)
-        data = await self._send_frame(frame, slave_id, function)
-        return RawModbusResponse(self._decode_register_words(data, count=count))
-
-    async def read_input_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        _ = attempt
-        self._validate_slave_id(slave_id)
-        self._validate_read_count(count)
-
-        async def _invoke() -> RawModbusResponse:
-            return await self._read_registers_common(
-                slave_id=slave_id,
-                address=address,
-                count=count,
-                function=4,
-            )
-
-        return await self._execute(_invoke, ensure_connection=True)
-
-    async def read_holding_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        count: int,
-        attempt: int = 1,
-    ) -> Any:
-        _ = attempt
-        self._validate_slave_id(slave_id)
-        self._validate_read_count(count)
-
-        async def _invoke() -> RawModbusResponse:
-            return await self._read_registers_common(
-                slave_id=slave_id,
-                address=address,
-                count=count,
-                function=3,
-            )
-
-        return await self._execute(_invoke, ensure_connection=True)
-
-    async def write_register(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        value: int,
-        attempt: int = 1,
-    ) -> Any:
-        _ = attempt
-        self._validate_slave_id(slave_id)
-
-        async def _invoke() -> RawModbusWriteResponse:
-            frame = self._build_write_single_frame(slave_id, address, value)
-            response = await self._send_frame(frame, slave_id, 6)
-            self._validate_write_echo(response, address=address, expected_value=(value & 0xFFFF))
-            return RawModbusWriteResponse()
-
-        return await self._execute(_invoke, ensure_connection=True)
-
-    async def write_registers(
-        self,
-        slave_id: int,
-        address: int,
-        *,
-        values: list[int],
-        attempt: int = 1,
-    ) -> Any:
-        _ = attempt
-        self._validate_slave_id(slave_id)
-        self._validate_write_count(len(values))
-
-        async def _invoke() -> RawModbusWriteResponse:
-            frame = self._build_write_multiple_frame(slave_id, address, values)
-            response = await self._send_frame(frame, slave_id, 16)
-            self._validate_write_echo(response, address=address, expected_value=len(values))
-            return RawModbusWriteResponse()
-
-        return await self._execute(_invoke, ensure_connection=True)
+from .modbus_transport_client import (
+    SERIAL_IMPORT_ERROR,
+    RtuModbusTransport,
+    TcpModbusTransport,
+    _AsyncModbusSerialClient,
+    _ClientBackedTransport,
+)
+from .modbus_transport_raw import RawRtuOverTcpTransport
+
+__all__ = [
+    "BaseModbusTransport",
+    "RawModbusResponse",
+    "RawModbusWriteResponse",
+    "RawRtuOverTcpTransport",
+    "RtuModbusTransport",
+    "SERIAL_IMPORT_ERROR",
+    "TcpModbusTransport",
+    "_AsyncModbusSerialClient",
+    "_ClientBackedTransport",
+    "_append_crc",
+    "_crc16",
+    "classify_transport_error",
+    "get_rtu_framer",
+]
